requirements.txt
# Deep Learning Framework
torch>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment Tracking
wandb>=0.16.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Web Server
flask>=3.0.0

# SSL Libraries (Optional - depends on your implementation)
# lightly>=1.4.0  # SSL framework
# timm>=0.9.0     # Pre-trained models

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0